Handle existing databufs in paper_*_databuf_create
If the databuf to be create already exists and has the same size as is
being requeested, then attach to it and return the pointer rather than
erroring out.  This allows a single call to paper_*_databuf_create
rather than a call to create followed by a call to attach if the create
failed.