Convert paper_gpu_thread to loadable module
Loadable modules will be a new way of building pipelines at runtime.
This is the first step of converting a single thread (paper_gpu_thread)
to the new loadable module model.  The paper_xgpu program is also
modified to use the new paper_gpu_thread loadable module.

Future steps will convert other threads to loadable modules and to allow
for the pipeline modules to be given on the command line in the order in
which they should be connected.  This will allow a generic program to
build arbitrary (and possibly invalid) pipelines.