Do not #ifdef out functions in guppi_udp.[hc]
Some functions had been #ifdef'd out of guppi_udp.[hc] if SPEAD was not
defined.  These functions are not harmful if SPEAD is not defined and if
SPEAD is not defined then the lack of these functions breaks
vegas_net_thread.  For these reasons, these functions are now always
included in the headers and objects.