Add integration control to paper_gpu_thread.c
Integrations are controlled by a few status buffer variables:

INTSTAT is a string based control/status variable.  Setting it to
"start" via an external process will cause the GPU thread to start
integrating when it gets a data block that containing the mcount given
in the INTSYNC variable.  Once integration has started, INTSTAT will be
st to "on".  To stop integration, set INTSTAT to "stop".  When the
current integration finishes, integration will stop and INTSTAT will get
set to "off".

INTSYNC specifies the mcount at which integration will (or did) start
when INTSTAT is set to "start" (or "on").  Note that setting INTSYNC
alone will not start integrations.  The actual enabling of integrations
is controled by INTSTAT.

INTCOUNT is the number of mcounts per integration.  It is read whenever
INTSYNC transitions from "start" to "on".

INTSYNC and INTCOUNT must be multiples of N_SUB_BLOCKS_PER_INPUT_BLOCK,
but there are currently no checks or adjustments to enforce this.