Fix dangling pointer problem in hgetc()
hgetc() returned a pointer to a non-static local variable.  Non-static
local variables live on the stack and becomes invalid after the return.
Using the returned pointer in or after another function call could be
very bad.

The fix is to make sure that the returned pointer points to a static
local variable.  Static local variables do not live on the stack and
pointers to them remain valid.

This does NOT make hgetc thread-safe.  That would require either using
thread-local storage or requiring the caller to pass in a pointer to a
caller-managed buffer.