Restore some pre-autotools Makefile options
Some of the VEGAS options could probably be better implemented as
configure --with-XXX options, but the future of the VEAGS code in this
codebase is questionable anyway so this is a very low priority.  One
side effect of doing these options the "old" GNU make way of ifdef/endif
blocks is that "endif" needs to be handled carefully to avoid confusing
automake.  Basically, "endif" tokens intended for GNU make get
AC_SUBST'd as @gnumake_endif@ so that automake doesn't see them.