Add index functions for various data orderings
Add functions to compute the index of the real component of the cross
product of two given inputs.  One function, regtile_index, computes the
index for data stored in "register tile order", which is the order
output by the xGPU code.  The other function, casper_index, computes the
index for data stored in "casper order", which is the order expected by
the dowsnstream data catcher (cn_rx.py).