Added compile options to enable debug test pattern
Allows the use of FAKE_TEST_INPUT and, optionally, FAKE_TEST_CHAN and
FAKE_TEST_INPUT1 to be defined to set input FAKE_TEST_INPUT (and
optionally input FAKE_TEST_INPUT1) to -16 for channel FAKE_TEST_CHAN
(defaulting to channel 0 if FAKE_TEST_CHAN is not defined).  All other
inputs and channels are set to 0.  Note that FAKE_TEST_INPUT must be
defined for the other optional values to have effect.