First, rough, draft of the conversion of the unpacking code from the use if indexes to the use of pointer arithmetic, with redundant arithmetic factored out.
This code is untested!

There is an mcnt related bug that crops up when the code is run
against the network rather than in timing test mode.

The input buffer structure hierarchy was changed and this broke the
builds of paper_gpu_thread and guppi_fake_net_thread.  This is not
fixed in the current commit.