Add compile time option COALESCE_FLUFF
Defining COALESCE_FLUFF when compiling paper_net_thread.c will perform
the unpack and fluff at the same time.  Ironically, this appears to be
slower than unpacking a complete block, then fluffing the complete
packet.  I think this is due to improved cache efficiency when doing the
fluffing on the full block.