Fix cache alignment issue in paper_input_databuf_t
Now that we are accounting for the size of the struct guppi_databuf at
the beginning of paper_input_databuf_t, we need to ensure that the data
in the paper_input_databuf_t that follow the guppi_databuf member are
cache aligned by adding some padding after the guppi_databuf member.  We
also must account for that padding when calculating pointers into the
input buffer.