Change the scaling of mcnt sent to the catcher
The mcnt scaling factor of the ROACH F engine is different from that of
the old iBob F engine which is different from that of the old ROACH X
engine.  The mcnt sent to the catcher is now scaled to remain compatible
with scaling factor output by the old X engine.  This allows the same
catcher software to be used with both FPGA and GPU X engines.