Make guppi_status_monitor.py "instance-aware"
guppi_status_monitor.py can now be given a status buffer instance_id on
the command line.  While it it running, 0-9 can be pressed to select
instance 0 through 9.  Additionally, '-' and '+' (or '=') can be used to
select the previous or next status buffer instance.  The currently
displayed status buffer is shown in the top line.