Change scaling of FID to match new gateware
The FID field in the packet header used to be the first dual pol antenna
index for the given F engine.  Given 8 inputs (4 dual pol antennas) per
F engine means that the FID field used be a multiple of 4.

Now the FID field is truly an F engine index, so to calculate offsets
for a given F engine, the FID is now multiplied by N_INPUTS_PER_FENGINE
(i.e. 8) rather than 2.