Track recent mcnts and detect databuf errors
The most recent million mcnts seen from all packets is stored in a ring
buffer.  If databufs are marked as filled out-of-sequence, then an
assortment of debugging info is prented on stdout, the most reent mcnts
are written to mcnt.log, and the program aborts (and generates a core
file if ulimit permits).