Change when paper_input_databufs are marked filled
The code used to wait for a block to be completely full, then it
would mark the block two before the current block as filled.  This is
problematic when packets are lost because incomplete blocks would mean
that two blocks before the incomplete block would never get marked as
filled.

The code now marks the block two before the current block as filled on
the first packet of the newly current block.  This is not the be all end
all solution because it doesn't handle the case where more than one
block's worth of packets are lost.  That is handled by the out of
sequence checking, which might have similiar (not yet resolved?) issues.