Update NETMCNT only when marking a block filled
Fixes a bug that was inadvertently introduced in commit 0fc570f.  That
commit changed when blocks are marked as filled, but it inadvertently
also changed the return sematics of write_paper_packet_to_blocks() in
such a way that the NETMCNT status variable was updated every packet
instead of just when marking a block as full.  This had a very
detrimental impact on packet processing times.