Do not update status on every packet received
Since commenting out (and removing) blocking network recevies, the
"waiting" variable in paper_net_thread was never set to 0, so the status
would be set to "WAITING" on *every* packet recevied!  The waiting
variable and the over agressive status updating has been removed.  The
status of the network thread is simply set to "running" upon entering
the run() function.