Compute net thread throughput and store in status
This calls clock_gettime before calling recv and again after the packet
has been handled.  The differnce is accumulated and packets are counted.
Once a block is sent off, the average number of nanoseconds per packet
is calculated.  The number of bits per packet divided by the average
number of nanoseconds per packet gives the average Gbps of the network
thread.  Note that this since the net thread uses non-blocking sockets
and clock_gettime is called before each call to recv, the computed Gbps
does NOT include time spent waiting for packets so it is a measure of
the network thread's throughput rather than a measure of the actual
data rate.  Needless to say, the throughput must be greater than (or
equal to) the data rate in order to keep up.