Allow using un-installed xGPU
When installed, the required xGPU files are in
${XGPUDIR}/{bin,include,lib}, but prior to installed, they all live in
the same directory.  This change first looks for the required files in
the bin, include, and lib subdirectories of ${XGPUDIR} and, if not found
there, in the ${XGPUDIR} directory itself.  This allows one to build and
run using a local, un-installed version of xGPU.