Change fluffing and GPU input buffer ordering
GPU input order is now:

  +--time--+      +--chan--+      +--fid---+
  |        |      |        |      |        |
  V        V      V        V      V        V
  m        t      x        c      q        f
  m        t      c        x      q        f
  ==      ==      ==      ==      ==      ==
  m0 }--> t0 }--> c0 }--> x0 }--> q0 }--> f0
  m1      t1      c1      x1      q1      f1
  m2      t2      c2              q2      f2
  :       :       :               :       :
  ==      ==      ==      ==      ==      ==
  Nm      Nt      Nc      Nx      Nq      Nf

instead of:

  +--time--+      +--chan--+      +-inputs-+
  |        |      |        |      |        |
  V        V      V        V      V        V
  m        t      x        c      q        f
  ==      ==      ==      ==      ==      ==
  m0 }--> t0 }--> x0 }--> c0 }--> q0 }--> f0
  m1      t1      x1      c1      q1      f1
  m2      t2              c2      q2      f2
  :       :               :       :       :
  ==      ==      ==      ==      ==      ==
  Nm      Nt      Nx      Nc      Nq      Nf

The "x" and "c" indices have been swapped.  This provides for a fixed
stride for each word in a packet, which allows for more efficient
fluffing, including the use of "round robin" fluffing (i.e. fluff four
packets at a time) rather than "sequential" fluffing (i.e. fluff one
packet at a time).

One side effect of this is that channels for different XIDs will be
interleaved going in and out of the GPU.