Time three distinct states of paper_net_thread
The three states that are timed are "wait" (how long do we wait before
getting another packet), "recevie" (how long does the recv call take),
and "process" (how long do we take to process the packet).  The sum of
receive and process, i.e. the total time we take to handle a packet, is
used to calculate paper_net_thread's throughput.