Make die on out of seq error a compile time option
In production, marking blocks filled out of sequence (a serious error)
should not result in terminating the process.  These errors should
"never" happen, but robust handling of them is critical if the
"impossible" happens.

In development, it may be desirable to end the process if this occurs
(e.g. it's a convenient way to stop execution in the debugger) so this
behavior cn be enabled by compiling with "DIE_ON_OUT_OF_SEQ_FILL"
defined (e.g. "make CFLAGS=-DDIE_ON_OUT_OF_SEQ_FILL").