Compute block to send from N_INPUT_BLOCKS, not 2
The block to send upon marking block B filled was computed as "(B-2) mod
N_INPUT_BLOCKS" (i.e. two blocks ago).  This is not suitable for when
N_INPUT_BLOCKS is large.  We now compute the block to send as
N_INPUT_BLOCKS/2 block ago.  This may not be ideal, but at least it
adapts dynamically to the value of N_INPUT_BLOCKS.