Add debug support for tracking missed packets
If paper_net_thread.c is compiled with N_DEBUG_INPUT_BLOCKS defined as
1, an extra block is included in the paper_input_databuf.  This extra
block is used to log the headers of all incoming packets (wrapping
around to the beginning of the block when the end is reached).  If a
packet is missed, the next 65536 packet headers are logged, then the
program exits so that the packet headers can be examined (e.g. with
dump_databuf.c) to see which packets were missed.