Always compute keys in guppi_{databuf,status}_key
The computed keys used to be stored in a static variable inside these
functions using lazy initialization.  This means that they will always
return the key of the first instance_id requested regardless of the
value of the instance_id parameter on subsequent calls.  This means that
a single process would only be able to compute a single key for each
buffer type.  This is bad!

Computing the key each time fixes this problem.