Make guppi_databuf_{create,clear} generic
This removes the GUPPI specific code from guppi_databuf_create and
guppi_databuf_clear.  It also simplifies the structure of a
guppi_databuf based shared memory block.  The shared memory segment is
now structured as a header of user-specified size (but must be at least
large enough to hold a "struct guppi_databuf") plus a user-specified
number of blocks of a user-specified size.  The beginning of the header
will hold a "struct guppi_databuf".