Make hashpipe_status_detach tolerant of NULL
Being able to pass a pointer to a hashpipe_status_t structure with a
NULL buf field to hashpipe_status_detach makes pthread cleanup
management easier since pthread_cleanup_push must be in the same lexical
block ('{' to '}') as the corresponding pthread_cleanup_pop.

Passing hashpipe_status_detach a NULL pointer or a non-NULL pointer
tthat points to a hashpipe_status_t structure with a NULL buf field is a
no-op that always succeeds.