Rework Makefile.am to change build products
Makefile.am now builds the following libraries:

libhashpipestatus - Contains functions for interfacing with hashpipe
                    status buffer.

libhashpipe - Contains functions for application threads to interface
              with hashpipe.

paper_gpu - Contains application threads for PAPER GPU-based X engine.
	    This is built as a hashpipe plugin.  The hashpipe executable
	    does not yet load plugins, but it will soon.

Makefile.am also builds the following executables:

check_hashpipe_databuf - This is a crude utility for examining hashpipe
                         databuffers.

check_hashpipe_status - This is a crude utility for examining and
                        manipulating the hashpipe status buffer.

clean_hashpipe_shmem - This is a crude utility for "cleaning" hashpipe
                       shared memory.

dump_hashpipe_databuf - This is a crude utility for dumping a hashpipe
                        databuf.

hashpipe - This is the main hashpipe executable.