More preparations for plugin architecture
Eliminating the use of the term "module" and changing the names of some
typedefs and structures to be more descriptive of theie purpose.

Moved some functions into a new file, hashpipe_plugin.c, but those will
probably be moved (back) into hashpipe_thread.c and the functions
currently in hashpipe_thread.c will be moved to hashpipe.c (and removed
from the paper threads).