Force including rpath $(libdir)
It seems that the autotools does not always include rpath=$(libdir) in
the executable (maybe if libdir is a directory included in ld.so.conf?).
It also seems that ldconfig will not cache the plugin shared libraries
(maybe because they don't start with "lib..."?).  To make hashpipe able
to find plugins installed into $(libdir), Makefile.am now explicitly
includes language that makes an rpath entry in the executable for
$(libdir).