Make cast in pthread_cleanup_push calls better
Change how the cleanup callback is cast when passing a pointer to it to
pthread_cleanup_push.  It used to be case as a (void *), but that was
not pedantically correct and sometimes the compilation of client code
would complain about this cast in the hashpipe_status_lock_safe macro.

The cast is now done to a (void (*)(void *)) which matches the
declaration of pthread_cleanup_push.