Install hashpipe setuid and run inits as root
Hashpipe is now installed with the setuid permission bit set.  When
installed as root, this means that the process starts with root
privileges.  The program temporarily drops root privileges when it loads
plugins, then restores previleges while it runs the threads' init
functions, and finally drops root privileges before calling the threads'
run function.  Essentially, this means that the thread init functions
run with root privileges.

The primary motivation for this is to simplify the use of packet
sockets, which require root privileges to setup.  Ideally the program
would simply drop all capabilities except for the CAP_NET_RAW
capability, but that refinement will have to wait for another day.

One details about being setuid is that any shared libraries must be
found in directories that are in ld.so.conf (or ld.so.conf.d).  This
includes Hashpipe plugins and any shared libraries they require (e.g.
CUDA libraries).