Add long missing `break` statement to getopt case
The 'p' case of the getopt switch statement has been missing a break
statement for a long time.  This went undetected for so long because the
'p' case "falls throught" to a non-implemented 'b' case that does
nothing except break.  Once the 'V' case was added between the 'p' and
'b' cases, specifying '-p' effectively also specified 'V' which prints
the version and exits!