Force "header" keywords to 8 character limit
The "header" (e.g. status buffer) keywords that were actually stored in
memory were already limited to 8 characeters by hputc.  This
functionality was inherited from the primordial code base.  This leads
to problems if a keyword with more than 8 characters is "put" into the
header more than once since the longer keyword on subsequent puts will
not match the truncated keyword from previous puts thereby adding the
keyword/value pair to the header in multiple locations.  To prevent
this, keywords are now explicitly limited to 8 characers in hputc and
ksearch.