Use more precision in hputr4/hputr8
The hputr4 and hputr8 functions converted the floating point value to a
string using "%f" and "%g", which default to only 6 digits of precision.
The max number of decimal digits is 9 for single precision and 17 for
double precision so to avoid loss of precision, we now use "%.9f" and
"%.17g".