The -o option now auto-recognizes types of values
When using `-o KEYWORD=VALUE`, VALUE is first converted to an integer
(using `strol`).  If that consumes all the charachers, the integer value
is stored in the status buffer, otherwise VALUE is converted to a
double (using `strtod`).  If that consumes all the characters, the
double value is stored in the status buffer, otherwise VALUE us stored
as a string.