hashpipe_irqaffinity.sh retries if no IRQs found
If no IRQs are found when setting the IRG affinity,
hashpipe_irqaffinity.sh will sleep for a millisecond and then check
again.  This will happen up to 2000 times (i.e. 2 seconds) befoire
giving up.  This is to handle the unlikely case of the interrupts not
being fully setup when the script starts.