Improve the way send work completions are handled
This involves swapping the wr_id values from the head and tail elements
when posting send work requests.  When processing work completions, this
allows the linked list of work requests to be pushed back onto the head
of the list of idle send packets.

A utility function was also created and some debug functionality can be
optionally compiled in by defining VERBOSE_IBV_DEBUG when compiling
hashpipe_ibverbs.c.