Improve completion notifications for send WRs
Previously, IBVerbs queue pairs (QPs) were created with `sq_sig_all` set
to 0.  When sending a linked wist of send work requests (send WRs), the
final WR of the list would get `IBV_SEND_SIGNALED` bitwise-or'd into its
`send_flags` field.  The thinking was that this would cause work
completions (WCs) to be posted to the send completion queue (send CQ) en
masse for the entire list of WRs upon completion of the final WR.  Such
thinking was flawed.  It turns out that when the QP has `sq_sig_all` set
to 0, only WRs with `IBV_SEND_SIGNALED` set in `send_flags` would get
WCs posted to the send CQ.  This meant that only the last WR of a sent
linked list of WRs would be "recycled".  The other WRs of the linked
list without `IBV_SEND_SIGNALED` would never be detected as "completed"
and would never be reused!

To fix this problem, the QPs are now created with `sq_sig_all` set to 1
and the `IBV_SEND_SIGNALED` flag is never explicitly set or cleared from
send WRs.