Add UDP sniffer flow diagnostic to ibvpkt thread
The special IB verbs "sniffer flow" (enabled by `IBVSNIFF`) does not
"consume" the matched packets, but instead copies the packets to the
work requests *and* allows the packets to pass through to the kernel.
The new "UDP sniffer flow" is a regular UDP flow that matches the MAC
address of the interface and a destination UDP port.  Because it
currently matches on the interface's (unicast) MAC address, it won't
match UDP packets sent to the broadcast or multicast MAC addresses (for
now).  This UDP sniffer flow is stored in the ibv_context's last flow
index.  It is enabled by storing `UDPSNIFF=<port>` in the status buffer,
where `<port>` is the desired destination UDP port.  Setting it to `0`
will stop the flow.  Setting it to `-1` will disable this altogether.